requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
selenium==4.16.0
lxml==5.1.0
python-telegram-bot==20.7
PyPDF2==3.0.1
pdfplumber==0.11.0
